+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## NEXT
+
+- Made `cai_causal_graph.causal_graph.CausalGraph.copy` more general to preserve the class type. Removed `.copy` from
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` class.
+
 ## 0.2.5
 
 - Added the `cai_causal_graph.identify_utils.identify_confounders` utility function, which allows you to identify